[#82] Analytic utils && handlers (#92)
[#82] Analytic utils && handlers

+++ jiraConnector/models/transformerModel.go
@@ -15,5 +15,5 @@ type TransformedIssue struct {
 	CreatedTime time.Time
 	ClosedTime  time.Time
 	UpdatedTime time.Time
-	Timespent   int
+	Timespent   int64
 }
